+++ extension/frontend/src/api/deleteDownload.ts
@@ -1,0 +1,7 @@
+import { axiosInstance } from '@/api/axios.ts';
+
+export default async (serverUuid: string, jobId: string): Promise<void> => {
+  await axiosInstance.delete(
+    `/api/client/servers/${serverUuid}/calaworkshop/downloads/${jobId}`,
+  );
+};